stability_index_computation optimization (#318)
* Push initial optimization for stability_index_computation

* Add persist and persist arguements and remove immediate path

* Remove unnecessary arguments and clean up workflow

* Add unit test for stability_index_computation

* Change to match kurtosis logic

* Add appended_metric_path back to stability_index_computation

* Add binary column and appropriate changes to workflow and report

* Remove unnecessary arguments in stability test

* Editable install instead of PATH hack in conftest

* Address all review comments, rename argument, modules, and add existing_metric_path

* Change appropriately in workflow and rename module in unit test

* Address comment to remove duplicate code

* Remove extendable install

* Add import back to conftest.py

* Add editable anovos install and remove import from conftest.py

* Change in setup.py to reflect module name change

* Update workflow.py

* Format drift_stability/validations.py using black

* Change to follow drift main branch structure
