+++ config/configs_full.yaml
@@ -233,7 +233,6 @@ drift_detector:
         mean: 0.5
         stddev: 0.3
         kurtosis: 0.2
-      existing_metric_path: ''
       appended_metric_path: 'si_metrics'
       threshold: 2
